Option to save full sized image
The level set output, by default, is the small VOI used for processing.  If the
"-f" or "--fulloutput" flag is set, save the level set as a full sized image of
the same size is the input.  This option is useful for processes that may not
honor origin and/or spacing in NIFTI files.

Add boolean option to GenerateLesionSegmentation.xml and logic to
GenerateLesionSegmentation.cxx.