ENH: multiple enhancements (see details)
1. Use external ITK (affine) transform file as a parameter (-t) to
initialize the pose of model
  - please note that the pose fitting includes only rigid + scale so
other potential transform in the file (if any, e.g. shearing) would be
reversed during fitting

2. Support the distance transform image as the input for fitting (-e
DT). This is to enable the bootstrapping strategy we discussed in the
meeting. With this mode, the current ASM could be used to find the
(rough, initial, or final?) correspondence from the distance transform
of a new sample image. An example usage in this mode is:

FitShapeModel -i
/Users/jinho/temp/10017X_INSP_STD_BWH_COPD_leftLung-128_regDT.nrrd -d
/Users/jinho/temp/output -o /Users/jinho/temp/output-temp-dt.nrrd -g
/Users/jinho/temp/output-geom-dt.asc -t /Users/jinho/temp/temp.tfm -r
Fitting -e DT -l 100 -m 200 -n 100

The output point file will be saved in output-geom-dt.asc, other
extensions to save the output geometry as a text point file (instead of
mesh) are: PTS, PNT, TXT (all same format)

3. Use the internal state of iso-surface of Poisson meshing to create
binary image of the fitted point model directly (i.e. without going
through mesh). With this method, I was able to remove the artifact of
the ITK TriangleMeshToBinaryImageFilter (a.k.a. missing lines) entirely
without additional cost. This mode is enabled with (-r Segmentation)
option.