ENH: add capability to create surface model with combined lung ASM
When -segmentation flag is used, the fitted lung geometry is created in
OBJ format using Poisson Surface Reconstruction. When combined lung ASM
is used, this Poisson method would not work correctly with the combined
point cloud. So the Poisson method is used separately for left and
right part of lung and combined later in the final output.